requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cloudinary==1.26.0
confusable-homoglyphs==3.2.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==4.1
django-bootstrap4==3.0.1
django-crispy-forms==1.13.0
django_heroku==0.3.1
django-jazzmin==2.4.9
django-media-pil==1.0.0
django-registration==3.1.2
django-registration-redux==2.9
django-tinymce==3.3.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==2.1.2
lts==0.0.2
Pillow==8.4.0
pycparser==2.21
pycrop==1.0.1
python-decouple==3.5
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
